requirements.txt
flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pokemontcgsdk==3.4.0
flask==3.0.0
flask-migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0